+++ schemacrawler-api/src/main/java/schemacrawler/crawl/MutableTable.java
@@ -27,6 +27,7 @@ import java.util.Set;
 import schemacrawler.schema.Column;
 import schemacrawler.schema.ColumnReference;
 import schemacrawler.schema.DatabaseObject;
+import schemacrawler.schema.EntityType;
 import schemacrawler.schema.ForeignKey;
 import schemacrawler.schema.Index;
 import schemacrawler.schema.NamedObject;
@@ -63,7 +64,8 @@ class MutableTable extends AbstractDatabaseObject implements Table {
   private final Set<DatabaseObject> usedByObjects = new HashSet<>();
   private MutablePrimaryKey primaryKey;
   private int sortIndex;
-  private TableType tableType = TableType.UNKNOWN; // Default value
+  private TableType tableType;
+  private EntityType entityType;
   private boolean isSelfReferencing;
   private String definition;
 
@@ -71,7 +73,10 @@ class MutableTable extends AbstractDatabaseObject implements Table {
     super(schema, name);
     requireNotBlank(name, "No table name provided");
 
+    // Set default values
     definition = "";
+    tableType = TableType.UNKNOWN;
+    entityType = EntityType.unknown;
   }
 
   /**
@@ -97,6 +102,12 @@ class MutableTable extends AbstractDatabaseObject implements Table {
     }
 
     return comparison;
+  }
+
+  /** {@inheritDoc} */
+  @Override
+  public EntityType getEntityType() {
+    return entityType;
   }
 
   /** {@inheritDoc} */
@@ -370,6 +381,12 @@ class MutableTable extends AbstractDatabaseObject implements Table {
   final void setDefinition(final String definition) {
     if (!hasDefinition() && !isBlank(definition)) {
       this.definition = definition;
+    }
+  }
+
+  final void setEntityType(final EntityType entityType) {
+    if (entityType != null) {
+      this.entityType = entityType;
     }
   }